backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "taxini-backend"
version = "0.1.0"
description = "Taxini Backend API built with FastAPI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Taxini Team" }]
keywords = ["fastapi", "backend", "taxi", "ride-hailing", "uv"]
dependencies = [
    "fastapi[standard]>=0.111",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "SQLAlchemy>=2.0",
    "sqlmodel>=0.0.24",
    "alembic>=1.16.4",
    "supabase>=2.18.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "pyjwt>=2.10.1",
    "gotrue>=2.0.0",
]

[project.optional-dependencies]
postgres = ["psycopg[binary]>=3.1", "asyncpg>=0.28"]
mysql = ["pymysql>=1.1", "aiomysql>=0.2"]
redis = ["redis>=5.0"]
auth = ["python-jose[cryptography]>=3.3", "passlib[bcrypt]>=1.7"]
s3 = ["boto3>=1.26"]

[project.urls]
Homepage = "https://example.com/taxini"
Repository = "https://example.com/taxini/taxini-backend"
Documentation = "https://example.com/taxini/docs"
Issues = "https://example.com/taxini/taxini-backend/issues"


[tool.hatch.build.targets.wheel]
# src/ layout: install the main package from src/
packages = ["src/"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "ruff>=0.5.0",
    "black>=24.4.0",
    "mypy>=1.8",
    "pre-commit>=3.7",
    "ipython>=8.18",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "httpx>=0.28.1",
]
test = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "coverage[toml]>=7.4",
    "pytest-cov>=5.0",
]
db = ["alembic>=1.13", "sqlalchemy-utils>=0.41"]

[tool.uv]
# Ensure the project package is installed into the environment on sync
package = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PL", "RUF"]
ignore = [
    "D", # pydocstyle checks are disabled by default; enable later if desired
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["taxini"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
line-ending = "lf"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --disable-warnings"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src/"]

[tool.coverage.report]
show_missing = true
skip_covered = true
precision = 2

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
exclude = ["tests/"]
